src/gpu_router.py
#!/usr/bin/env python3
"""
GPU Router Engine - Core GPU-accelerated routing algorithms
"""

import cupy as cp
import numpy as np
import time
from dataclasses import dataclass
from typing import List, Tuple, Optional
import logging

logger = logging.getLogger(__name__)

@dataclass
class RoutingResult:
    """Result of routing operation"""
    success: bool
    tracks_created: int = 0
    success_rate: float = 0.0
    error: Optional[str] = None
    routing_time: float = 0.0

class GPUAutorouter:
    """GPU-accelerated PCB autorouter using CuPy/CUDA"""
    
    def __init__(self, board):
        """Initialize GPU autorouter with KiCad board"""
        self.board = board
        self.grid_size = 0.1  # mm
        self.logger = logging.getLogger(__name__)
        
        # Check CUDA availability
        try:
            self.device = cp.cuda.Device(0)
            self.logger.info(f"✅ GPU initialized: {self.device}")
        except Exception as e:
            self.logger.error(f"❌ GPU initialization failed: {e}")
            raise
    
    def route_board(self) -> RoutingResult:
        """Main routing function"""
        start_time = time.time()
        
        try:
            self.logger.info("🎯 Starting GPU-accelerated routing...")
            
            # Get board dimensions and create grid
            bounds = self._get_board_bounds()
            grid = self._create_routing_grid(bounds)
            
            # Get nets to route
            nets = self._get_unrouted_nets()
            self.logger.info(f"Found {len(nets)} nets to route")
            
            if not nets:
                return RoutingResult(
                    success=True,
                    tracks_created=0,
                    success_rate=1.0,
                    routing_time=time.time() - start_time
                )
            
            # Route each net using GPU acceleration
            routed_count = 0
            for i, net in enumerate(nets):
                if self._route_net_gpu(net, grid):
                    routed_count += 1
                
                # Progress update
                progress = (i + 1) / len(nets) * 100
                self.logger.info(f"Progress: {progress:.1f}% ({routed_count}/{i+1} routed)")
            
            success_rate = routed_count / len(nets)
            routing_time = time.time() - start_time
            
            self.logger.info(f"✅ Routing completed in {routing_time:.2f}s")
            
            return RoutingResult(
                success=True,
                tracks_created=routed_count,
                success_rate=success_rate,
                routing_time=routing_time
            )
            
        except Exception as e:
            return RoutingResult(
                success=False,
                error=str(e),
                routing_time=time.time() - start_time
            )
    
    def _get_board_bounds(self) -> Tuple[float, float, float, float]:
        """Get board bounding box"""
        # Placeholder - implement with actual KiCad board API
        return (0, 0, 100, 80)  # mm
    
    def _create_routing_grid(self, bounds) -> cp.ndarray:
        """Create routing grid on GPU"""
        x_min, y_min, x_max, y_max = bounds
        
        grid_x = int((x_max - x_min) / self.grid_size)
        grid_y = int((y_max - y_min) / self.grid_size)
        
        # Create grid on GPU
        grid = cp.zeros((grid_x, grid_y), dtype=cp.int32)
        
        self.logger.info(f"Created {grid_x}x{grid_y} routing grid on GPU")
        return grid
    
    def _get_unrouted_nets(self) -> List:
        """Get list of nets that need routing"""
        # Placeholder - implement with actual KiCad board API
        return []  # Return actual nets from board
    
    def _route_net_gpu(self, net, grid: cp.ndarray) -> bool:
        """Route a single net using GPU pathfinding"""
        try:
            # Placeholder for GPU pathfinding algorithm
            # This would implement A* or wavefront on GPU
            
            # For now, just simulate success
            time.sleep(0.01)  # Simulate routing time
            return True
            
        except Exception as e:
            self.logger.error(f"Failed to route net: {e}")
            return False
