src/requirements.txt
kicad-python>=0.4.0
cupy-cuda12x>=12.0.0
numpy>=1.21.0
matplotlib>=3.5.0
